chore: Improve error handling for student endpoints

+++ main.py
@@ -1,10 +1,11 @@
-from fastapi import FastAPI, Depends
 from dotenv import load_dotenv
-from sqlalchemy.orm import Session
+from fastapi import Depends, FastAPI, HTTPException
 from fastapi.middleware.cors import CORSMiddleware
-from model.student import Student
-from database import SessionLocal, engine, Base
+from sqlalchemy.orm import Session
+
+from database import SessionLocal
 from dto.student import StudentDTO
+from model.student import Student
 
 load_dotenv()
 
@@ -38,7 +39,12 @@ async def get_students(db: Session = Depends(get_db)):
 
 @app.get("/students/{student_id}")
 async def get_student_by_id(student_id: int, db: Session = Depends(get_db)):
-    return db.query(Student).filter(Student.id == student_id).first()
+    student = db.query(Student).filter(Student.id == student_id).first()
+    
+    if not student:
+        raise HTTPException(status_code=404, detail="Student not found")
+    
+    return student
 
 
 @app.post("/students")
@@ -58,6 +64,10 @@ async def create_student(student_dto: StudentDTO, db: Session = Depends(get_db))
 @app.delete("/students/{student_id}")
 async def delete_student_by_id(student_id: int, db: Session = Depends(get_db)):
     student = db.query(Student).filter(Student.id == student_id).first()
+
+    if not student:
+        raise HTTPException(status_code=404, detail="Student not found")
+
     db.delete(student)
     db.commit()
     return {'message': 'Student deleted successfully'}
@@ -66,6 +76,10 @@ async def delete_student_by_id(student_id: int, db: Session = Depends(get_db)):
 @app.patch("/students/{student_id}")
 async def edit_student(student_id: int, student_dto: StudentDTO, db: Session = Depends(get_db)):
     student_to_edit = db.query(Student).filter(Student.id == student_id).first()
+    
+    if not student_to_edit:
+        raise HTTPException(status_code=404, detail="Student not found")
+
     student_to_edit.firstname = student_dto.firstname
     student_to_edit.lastname = student_dto.lastname
     student_to_edit.student_id = student_dto.student_id
